refactor: passenger actions now have they own directory
