+++ server/src/http/actions/passengers/update.passenger.action.ts
@@ -1,16 +1,17 @@
 import { Request, Response } from 'express';
-import { UpdatePassengerCommand } from '../../application/commands/update.passenger.command';
-import updatePassengerHandler from '../../application/handlers/passengers/update.passenger.handler';
+import { UpdatePassengerCommand } from '../../../application/commands/update.passenger.command';
+import updatePassengerHandler from '../../../application/handlers/passengers/update.passenger.handler';
+
 class UpdatePassengerAction {
   async run(req: Request, res: Response) {
     try {
       const command = new UpdatePassengerCommand(
         req.params.id,
-        req.body.fullName,
+        req.body.fullname,
         req.body.email,
         req.body.identityCard,
       );
-console.log(command)
+
       try {
         await updatePassengerHandler.execute(command);
       } catch (error) {
